+++ earth_extractor/providers/__init__.py
@@ -2,6 +2,5 @@
 from .base import Provider
 from .copernicus import CopernicusOpenAccessHub, copernicus_scihub
 from .alaskan_satellite_facility import asf
-from .sinergise import sinergise
 from .nasa import nasa_cmr
 from .swisstopo import swiss_topo